Revert "Rearrange the items to bring on an outdoor trip"
Reverting c7aa85cd1abb5a04e27982fc8c3e130a3dcccad8 so that
the list of items is set to the original order as in 11448c3ac19031172852628d630e6b5d381ffabb.

What the Revert command will do is it will take all of the changes that were there, and it will flip them around. It will do the exact opposite of those changes. So, anything that was added will be deleted, anything that was deleted will be added back in again, and anything that was modified will be changed back to its previous state. It's going to be a complete mirror image of this commit.